move to awk-like command syntax (#21)
The command line now expects at least one input file, which can
be specified as "-" for standard input.  Also, we allow the zql
expression to be optional by checking to see if the argument is
a valid file and interpret it as such instead of a search for
the string that represents the file.

Also, we added a brief description for help text.