fix bug in zeek _path output handling
For zeek output, we remove the _path column in each record and
use #path directives to indicate the _path value.  Previously,
we did not handle proper #path updates in the ouput stream when
comingling results with different path outputs in the same stream.
These changes fix this so that #path is updated when the _path
field changes.  If the path field disappears, then #path is
reset to empty by indicating "#path -".  If no _path exists
at the beginning of the stream, then no #path directive is
output at all.