add buffered writing for all writers (#36)
These changes introduce a wrapper package to transform a
bufio.Writer into an io.WriteCloser by calling flush on
close and arranges for all file output to be wrapped with
in this way.  Otherwise, every write on each file object
would result in a system call resulting in poor perfomance.