fix bugs parsing nested containers (#42)
These changes fix bugs in the way recursive containers were
handled.  The zval code was designed to handle legacy zeek logs,
that had at most one level of containers inside the top-level
record.  We refactored things into two passes where the first
pass unescapes the input values into an intermediary data structure
(which is allocated once per stream and reused across records),
and the second pass, computes the sizes of everything and
encodes the raw zval.

These changes also add the TBD code for unescaping escaped characters
in zson input. There is also now a check for a single "-" for a value
field, which now correctly turns into an unset column.

Finally, we are changing the zson spec to follow the zeek escaping
logic and refectored the zeek unescaping logic so it could be used
by the zson parser.