add reader/writer for zson over json
This commit implements readers and writers for sending
zson over json as ndjson.  This is a convenient format
for sending zson data to a native javascript client
(e.g., a web app).

Each value is sent as a JSON object with a field ("id") for
the descriptor id and a field ("value") that represents
an array of strings in zeek-escaped format.  The array
may contain sub-arrays of strings (and other sub-arrays)
if there are records inside of records, vectors, or sets.
If the id hasn't been transmitted previously in the stream,
then a third field ("type") encodes the zson type structure.