replace zson.Raw with zval.Encoding (#48)
These changes create a new type for zval encodings so that we
can distinguish between normal byte slices and byte slices that
are encoded zvals.  zson.Record.Raw becomes a zval.Encoding and
the zson.Raw type is deprecated.  zson.Record.Raw is still called
Raw as this is an apt name for the zval serialized encoding.