simplify interfaces to zson Writers
This commit better separates the concerns among the zson Writers.
The writers now have a Flush() interface method instead of Close()
and, when not needed, a simple wrapper provides a stub Flush() method.
We also added zson.WriteCloser which does the Flush on the WriteFlusher
when close is called.

These changes allowed us to refactor the emitter and simplify things.
There is now an emitter.Writer interface for both file and dir writing.

These changes also allow a simpler external reuse of the zson writers.

wip