add raw reader/writer to zsio (#50)
This commit adds support for reading and writing zson files
in a raw format, where each zson value is encoded directly
in its zval serialized encoding and each type descriptor binding
is encoded in its ascii form.  In both cases, a header frames
each line, where the header provides the type (descriptor, value, 
or zson comment), the descriptor id (if not a comment), and the
length of the buffer, and a virtual channel id, which is useful 
for tagging records with the output identifier for flowgraphs 
that have more than one output.

We also reserved a flag reserved to indicate that a value is
consists of machine encoded values instead of strings 
that need to be parsed.

We also added logic to avoid copying unmatched records for the
raw reader by lifting the filter into the scanner when possible.

This raw variant seems to go about 4X faster than zson scanning
for a simple count query, i.e., "* | count()".