change volatility default in zson.Records (#57)
In the previous commit, we changed the sense of the volatility
condition, which means zson.Record{} allocations would be
nonvolatile.  This commit reverse the logic and changes
the struct field from "volatile" to "nonvolatile" so that
zson.Record{} allocations are voltile by default.  When nonvoltile
records are desired, zson.NewRecord should be called.