remove limit from table writer (#71)
Instead of quitting, when the table writer reaches it limits, it
simply flushes its output and starts writing again.  This means
long outputs will periodically generate column headers.

The table writer does the same thing when the descriptor changes as
the header is no longerr consistent with the data.  So it flushes the
current data and writes a new header.