add autodetector for input formats
Previously, zq inferred the type of a file from its extension.
This is problematic especially since people often use the
same extension (e.g, ".log") for different types (e.g., zeek and ndjson).

This commit adds an autodetector that reads the first bit of each
file to figure out what type it is.