add support for control messages carried in zson.Record (#74)
This commit allows for control messages (aka zson comments)
to be carried in zson.Records and read/write to zson text and
raw formats.

Such records are not allowed into the proc flowgraph.

This mechanism is useful for higher layers to embed control
and signaling into a stream of raw zson records.  For example,
zson over http uses these messages to provide end-to-end
reliability semantics.