add channel to zson spec, misc refactor
This commit is a bit of a potpourri.  First, it adds the channel
abstraction to the zson spec and implementation so that channel IDs
can be attached each record both in the stream and internally.

Second, it revives the text writer so one can print output
the concise text format when perusing data.

Third, it refactors the emitter package to make it a bit easier
to re-use externally.