Add zeek.TypedEncoding and fix group-by mixed types (#84)
This commit adds zeek.TypedEncoding to make better sense of
zvals paired with zeek Types and it fixes a bug in group-by
where the different types encountered for the group-by key
fields would cause problems.

This commit got long because I kept running into zval bugs
knowing when/if the zval is a printable container or a unprintable
leaf container.  The String method on zeek.TypedEncoding now
takes care of thie logic and many places in the code that
previously carried around separate zeek.Type and []byte (zval)
is now simplified and easier to follow.

Also, we added zval.Encoding types many places where []byte
meant zval.Encoding.

Along the way, we added an Encode method to the zeek.Value interface
which makes it easier and less error-prone to build up the
zval.Encodings from zeek.Values.  This also gets us very to close
to supporting machine types in the zval raw format as we
merely need to change the way Encode works (and add a Decode method
to take the native zeek type and encode it as machine zval.Encoding)

Given the simpler model for value encoding, we also fixed cut to
encode the record zval directly and to properly handle variations
of the type of any given cut field.