fix problem with unset containers
The zeek parser was parsing unset and empty sets to the same zson.
We fixed this and realized the zson text format did not have
a way to represent an unset container.  So we added "*" to the
syntax to mean this like "-" means unset value for scalars.

Updated the parsing code to handle "*" and also updated
zson over json to encode unset containers as JSON empty object
inside of the JSON value arrays.  With the zval builder properly
handling empty container as empty container instead unset container,
a bug in ndjson was exposed where calling marshal on sets or vectors
didn't work right because the zeek value had its arrays unitialized.

Finally, updated zson spec.