fix problem with unset containers (#91)
The zeek parser was parsing unset and empty sets to the same zson.
We fixed this by adding methods to the zval builder to distinguish between 
container unset and container empty.  We fixed the zeek parser accordingly
then realized the zson text parser had the same problem and this was 
addressed by including type context, which we will need anyway when we 
convert strings to machine types.

We also updated zsio / ndjson accordingly.

These changes also exposed a bug in the zsio / ndjson code where 
zeek.Set and zeek.Vector were initialized with nil instead of []zeek.Value 
so the ndjson encoder improperly encoded empty values as unset.