Fix a bug with array index field expressions (#103)
When handling an out-of-bounds array index, compiled field expressions
were returning nil for both type and value, which is wrong.  In
particular, when groupby was grouping by an expression such as v[0],
if the first record it sees with a particular descriptor has a vector-valued
field v but the array index is out of bounds, groupby would erroneously
conclude that no records with that descriptor could ever match.
Fix the bug here by returning the proper inner type and a nil body
when an array index is out of bounds.