add shell script-driven testing
This commit provides a framework for tests run
as shell commands.  The test is defined with
input and expected output files the actual files
are checked against the expected file output.

An example (and useful test) is provided in
package tests/suite/diropt, which tests the
functionality of the -d and -o flags.