add shell script-driven testing (#108)
This commit provides a framework for tests run
as shell commands.  The test is defined along with
input files and expected output files.  After the run,
the actual files are checked against the expected file output.

An example (and useful test) is provided in package
tests/suite/diropt, which tests the functionality of
the -d and -o flags.