migrate zval encoding from strings to machine types (#124)
These changes simplify the zeek type system so scalar
values are representing directly as their native Go type
instead of having a struct that contained the Go value.

We also modified the CoerceTo functions so that they
can set a value by reference instead of creating a new one,
allowing the caller to allocate the value on the stack
thus avoiding sending temporary values through GC.

This commit makes a fundamental change from string rep to
machine-type rep for encoding values.

Each data type now has an Encode and Decode function to
convert to and from zval encoding to machine type.

The Type method New now takes a zval encoding instead of
a string.

A new zeek.Type method Parse transforms a human readable
string to the corresponding zval encoding for that type.

The Format method on zeek.Value was dropped as it is no
longer used.

These changes also get rid of zeek.None.  zeek.Unset is next.