auto-detect precision for zeek timestamps (#140)
This commit fixes the output format for zeek timestamps by
printing them with 6 fractional digits unless they are higher
precision, in which case they are printed with 9 fractional
digits.  We remember the high precision setting in the output
formatter so columns line up.  This means there can be a blip
in column alignment when changing from 6 to 9 digits but it
is stable thereafter. In practice, this should be rare and
the table output can be used when column alignment is important.

We also changed package nano to remove trailing zeroes from
its output to provide a more natural native representation for
time and duration given that zeek formatters no longer use
these methods.