Add -U flag for forcing output of utf-8 interpretation of zeek strings
This commit adds the -U flag for printing zeek string in a "forced"
UTF-8 format.  It is applicable to formats zeek, text, and table.

This is the very same non-standard formatting of mixed-binary strings
that zeek does.  There is no way to disambiguate between random binary data
and a deliberate utf-8 string so it's left to the "presenation layer" to decide
how to format the data.  This would be remedied if a zeek string were a UTF-8
string and zeek had a bytes type to represent non-string data.

For now, we have -U.