Preserve nested fields in groupby (#144)
* Split ColumnBuilder out of CutProc

The functionality added in PR #127 is also needed inside groupby,
split out the common bits so they can be re-used.

* Let zval.Builder handle encoded containers

For cases like "cut id" (where id is of type record[...]), we need
a zval.Builder to be able to append an already encoded container.
This is handled by adding another argument to Builder.Append() to
indicate that something is a container.

* Make groupby on fields inside records maintain key structure

Previously, something like "count() by foo.bar" would write the grouping
key as a field named "bar" in the output records.  With this change,
the structure is retained -- output records have a record called foo with
a single field bar.