rename ZSON spec to ZNG and rework a few things (#167)
These changes update the zson specification so it's now
called ZNG and reworked a few things. First, while originally
motivated to maintain backward compatibility with the zeek
log format, we decided to forego this constraints and
instead describe how zeek logs can be easily converted into
and out of ZNG format while preserving all of the original
zeek structure.

Also, we further developed the binary protocol and adopted
native machine data types a la protobufs and avro.  We also
developed the type system further so all types are represented
with a single integer type code and composite types are
built up from previous type code bindings.

The specification now "leads with" a description of the binary
format (called BZNG) and then describes the text format (ZNG)
in terms of the BZNG definitions.