Relax allowed characters in search terms, handle escaped = and *
Allow any characters to appear in search terms.  To clear up ambiguities
with characters like = and *, allow those characters to be escaped in
which case they are treated as part of the search term.  That is, the
query "foo=bar" is still interpreted as "match records in which the field
foo has the value bar"  ut the query "foo\=bar" is intepreted as "search
for the literal string foo=bar".  Additionally, don't treat escaped asterisk
characters as glob patterns.  So "foo*" matches any string that begins with
"foo", but "foo\*" matches any string that contains the sequence "foo*".