convert zng.Value to a struct from zng.TypedEncoding
These changes convert zng.Value from an interface to a
struct by replacing it with zng.TypedEncoding.  To do this,
we hang many of the old Value interface methods off of
zeek.Type so that the functionality remains and can now be
called using the pattern zng.Value.Type.Method(zng.Value.Bytes)

This is the first step toward eliminating value structs from
escaping to the heap in the inner loop of a scan.  (The next
step, in a susequent PR, will be to pool the Records and
the zcode.Bytes pieces so they can be easily recycled.)

These changes also elimate the unset type.  Instead an
unset value has a type and nil value and a nil has
has a nil tyoe and value.