Sort nulls (#200)
* Revamp sort argument handling

As sort gains additional arguments, having the grammar handle all
possible permutations of argument ordering is impractical.  Generalize
the grammar a bit so we don't have to enumerate all different orderings.

* Add -nulls parameter to sort

This parameter has two valid values, "first" and "last", specifying
where null values (ie the unset value or records that don't have the
named field) should be sorted first or last in the output.  The default
is "last".