Move zbuf.Record into package zng etc (#222)
This commit recreates the zng.Record value type based on the
zbuf.Record and deletes zbuf.Record.  It also merges zbuf.Descriptor
into zng.TypeRecord and eliminates the former.  This gets us
ready to create the generic type code and type context concepts for
the updated ZNG spec.

The zng writer also outputs in "canonical form", where each record
descriptor starts at id 0 and increases independent of the type IDs
in the internal stream.

resolver.Table is renamed resolver.Context to reflect its role as
the type context in the new ZNG spec and names throught are changed
to reflext "type context" over "descriptor table".

Unmarshaling of records has been updated to deal with the need to
create records and zng.Columns with a type context.  This is the
first move and will be imrpoved in a subsequent commit.

zbuf errors that were really zng errors have been moved to package zng.