fix bugs with marshaling zng.Context
This commit fixes bugs with the zng.Context and zng.Type marhaling
logic.  It also adds a test so marshaling problems may be detected
down the road since a regressions was previously missed.

Along the way here, we moved closer to supporting the latest ZNG spec
by introducing typedefs and type IDs.

Any external code that relies upon mashaling types will have to be
update as we are encouraging a model where you send a type as a
string conforming to the ZNG type system or you send ZNG directly
which embeds the type definitions, rather than marshaling the
Go type objects.

The marshaling logic for saving resolver.Contexts is a placeholder
until we get the ZNG implementation up to date.  At that point,
we will deprecate the marshal inteface on resolver.Context and
replace this with methods to read and write contexts as BZNG
files that exclusively contain typedefs.