Fix group-by keys bug and make type record allocation more robust (#225)
This commit fixes a group-by keys bug where a zero-length record
was illegaly put in the type context.  It also makes the creation
of new record types more robust by guaranteeing that the columns
of the newly created record have types (resurively speaking) that
all come from the context being called.  This way, LookupByColumns
cannot return a new record type that has columns with types that
point into othe contexts.

Along the way, we got rid of the stub type in ndjsonio, since any
old type would do (we changed it to string), when it calls unflatten
to find nested records.  With the change to LookupByColumns, the
stub type was no longer tolerated.

Finally, this commit also adds more friendly marshaling of
the type context table.  This code is all going away when we
switch to native BZNG files to represent type contexts, but it
was useful in debugging these problems here for now.