Derive reducer types from input types
We previously computed a reducer's result type from the first data is
saw, but we have cases (e.g., when the only records on which the reducer
operates have null values for the input field) where this doesn't work.
Fix this by augmenting the reducer interface to take the input record type
when it is instantiated and compute the result type from that.