use bzng instead of json to serialize resolver.Context type table
This commits changes the resolver.Context to use the BZNG data formrat
instead of JSON for serializing the type table comprisig the type context.
This a natural and more efficient way of storing the types used to
encode BZNG values, i.e., simply store the set of types as a BZNG
encoding of the types.

This commit also moves the resolver.File module to the bzngio package
now that type context files are stored as BZNG files, renaming
resolver.File as bzngio.TypeFile.

Finally, we noticed there was old, unused code for processing
deprecated bzng headers so this commit removes that dead code.