add stubs for zqd experiment (#280)
This is a drafty first cut at functionality for "zqd" a REST-based server for running zq queries.
It implements /search, /space, /space/:name, and /status. Currently, any subdirectory of the
directory that zqd is run in is included in the space list as long as it has a filed called "all.bzng".
The searches presented through /search are run on the corresponding all.bzng.