Separate unescaping paths for string/bstring
Unescape handling for the string type is not yet in place, but this is
an incremental step toward having separate unescaping rules in place for
the string and bstring types.  Also cleaned up several places where we
were double-unescaping unnecessarily (i.e., calling Unescape() and then
Parse() on the same data).