implement time ranges in /search and add /version
This commit properly limits a search to the requested time range.
It still scans the entire file and does not handle reverse searches
and it still ignore the dir option in the search request.

It also implements the /version endpoint, returning unknown for the
version info as the version number is not yet hooked up to the linker.

We also made the space info a bit more robust to computing the proper
time range (but still have no caching of this info so the entire
file is scanned on each info request).