create release assets for tags (#322)
When any tag matching a "v*" pattern is pushed to the repo, including
when a Github release is created, this action will create per-platform
zip files containing the binaries built from cmd/... . The zip files
will be attached to the Github release as release assets.

If the tag push was not due to a Release creation, this will create a
Github release and attach the zip files as assets. I'm not using an
action to explicitly create a release, as I didn't find anything yet
that would automatically pull the release changes text from the
changelog file; we can add that later.