test annotated examples in markdown docs (#301)
Add a test that combs markdown for annotated fenced code blocks. The doc
examples use zq-sample-data. Add a new dependency on test-system that
grabs the sample data if needed. The target is written in such a way
that it performs a one-time download but does not update it after that.

Doc authors can annotate their examples like so. What was once:

```
zq -f table '* | count()' *.log
```
```
COUNT
1234
```

Can now be:

```zq-command
zq -f table '* | count()' *.log
```
```zq-output
COUNT
1234
```

This is based on the CommonMark spec [0] which Github uses [1]. The
spec allows fenced code blocks to specify an info string [2]. In
practice this is used to denote the language. In our case we can parse
markdown and look for such strings to denote zq examples.

Goldmark [4] seemed to be the most suitable package for parsing
markdown, walking an AST, and finding fenced code blocks.

I've annotated all the examples that do not require record truncation.

[0] https://spec.commonmark.org/0.29/
[1] https://github.github.com/gfm/
[2] https://github.github.com/gfm/#example-113
[4] https://github.com/yuin/goldmark