support truncated output in markdown examples
Extend the markdown info string for zq-output to include an optional
'head:N' directive, where only the first N lines of output of a
zq-command are verified against its zq-output.

To find these we must extend our AST walk function. Rename ZQExamplePair
to ZQExampleInfo; add a new struct member to hold the number of lines to
compare. Do the same for ZQExampleTest.

Move the responsibility of actually running the zq-command to a
ZQExampleTest Run() method. This also handles any truncation.