condense main workflow into single job (#333)
Back in #14 when using btest it made sense to have a separate job, as a
different executor was needed. But since #87, when the executor was made
the same for both jobs, it no longer makes sense to have separate CI
jobs.

This removes a bit of complexity, the cost of job-to-job workspace
transfer, and 1 container initialization.