Comparison tests (#338)
Increase our coverage of all the different coerce-for-comparison
code paths in compare.go.  Also add one missing case (byte-float64
comparison) that the test uncovered.

Also remove CompareUint64() which was erroneously introduced in #325.
The CompareFoo methods are called on the type of the literal that
appears on the RHS of a filter with a comparison operator but there's
no way to create an unsigned literal in zql so this code is all unreachable.