add pcap indexing and slicing
This commit adds packages to do indexing and slicing of pcap files.
It currently supports only legacy pcap format (i.e., no support
yet for pcap-ng though it is designed so pcap-ng can be dropped in
without too much difficulty).

We also added a "pcap" command in cmd/pcap.  "pcap index" will create
a JSON index for use by "pcap slice".  "pcap slice" takes this
index, a time range, and a flow filter as arguments and produces
an extracted pcap file.  All of these arguments are optioonal.
If the index is provided, then the packets that fall outside of
the indexed time range are skipped, which dramatically speeds up the
slicing when exctracting a small range out of a large pcap.
If the time range is specified, it is used by the index and only
packets that fall within the time range are scanned.  (If the time
range is given but no index is provided, then all packets are scanned
but only packets that fall within the time range are matched.)
If a flow filter is specified in the format "ip:port ip:port",
along with a protocol ("tcp" or "udp"), then only packets from that
flow are matched.

The pcap packages are also used direclty by the pcap endpoint in zqd.
For now, the stiching is in package zqd/pcapdemo but this will change
in a subsequent PR when the new file handling model is in place.