Fix pcap slicer (#356)
The logic to read slices from an underlying pcap file was wrong as
io.SectionReader didn't work as expected.  This commit fixes
the Slicer to simply use seek on the underlying file to implement
the Read interface for the sections desired.

Also, fix bug where output pcap wasn't truncated when opening the 
file for writing.

And there was an off-by-one-nanosecond bug in the "-to" argument to 
"pcap slice" so we added nano.Span.ContainsClosed which checks if 
the timestamp is in the span as a closed-on-the-right interval.