Add space POST endpoint
Add the ability to create a new empty space via the space post
endpoint. The space post endpoint requires either a space name or
a space data dir (additionally both can be provided).

If a space data dir is specified, with no space name, a new space
is created given the name of data dir's base directory. If the name
collides with an exisiting space, a unique name is generated and
returned in the request response.