add support for ICMP pcap search
This commit adds support for searching for ICMP flows in pcaps.
It also adds hooks to "pcap slice" and the zqd packet end point
for filtering on icmp.

The commit also cleans up the pcap filtering API and internals by
creating more modular closures for packet filtering  At some
point, we may want to create a Go DSL for creating BPF programs
and port the BPF virtual machine and data flow optimizer to Go.
This would allow for much more flexible filtering, but is not
a priority right now.