Crude prototype for zng archive (zar)
This commit adds a zar command for createing and searching
bzng index files.  The index files are organized as sorted
string tables and use the sst package.

The current protoype is very early and and currently only
indexes IP addresses as a hard-wired configuration.  The code
however is set up to do more flexible indexing soon.

The sst package uses a simple binary format for framing
keys and values as arbitrary byte sequences but we realized
it could be a very powerful foundation to store the sst
files as bzng.  We will implement this change in a future PR.