zqd: atomic space config updates
Currently there could be an error when a concurrent request could be
reading from a space's config while another request is updating it.
The could cause the reader to read incomplete contents and therefore
get an invalid json error.

When updating a space's config file, first write to a temp file then
atomically rename the temp file to the standard `config.json` path.