WIP on expressions
Very much still in progress.  The grammar is mostly complete with the
exception of function calls and field/array dereferences (which are
going to require additional AST node types).  I've also begun to adopt
the convention in many other PEG grammars of using CamelCase rule names.
This means the grammar is in an inconsistent style right now but I
propose we go with this and clean up the rest of the grammar in a
separate pass when the rest of the work on expressions is finished.

Compilation is a different story, only addition of integers works right
now.  Making this work has required a tremendous amount of repetitive
code.  I haven't come up with any good ideas to simplify this yet,
definitely looking for ideas here...