create pcapio and generic Reader interface (#398)
This commit refactors the pcap package to get it ready to drop in
support for pcap-ng.  We moved the pcap reader implementation to
a new pcapio package that exports a generic interface for reading
pcap-ng blocks (section headers, packet blocks, inteface blocks)
in a way that is backward compatible with legacy pcap (a single
section header representing the pcap file header followed by
packet blocks).

All of the pcap parsing code imported from pcapgo will live in
the pcapio package, named with by their original source file names.

We also simplified the Index type since the struct contained
just a slice of Sections so now pcap.Index should be passed around
instead of *pcap.Index.  This also means that the marshaling of
this data structure changes the format of the on-disk index
file so existing spaces need a new run of "pcap index".

This commit is the second step toward getting the pcap subsystem
read for pcap-ng.  The next PR in this sequence will have the
actual support for pcap-ng.