Cache min and max times (#388)
* Add driver.NewDemuxed()

driver.NewDemuxed() takes a slice of zbuf.Writers which must be of
same length as the number of leafs in the flowgraph that is then run
on it. This allows clients to distinguish leaves whereas driver.New()
multiplexes all records into a same zbuf.Writer.

* Populate space min/max at ingest time

min and max times are pulled out when running the ingest zq flowgraph
on the freshly emitted zeek logs and are written into the data dir in a
file named info.json.