Progressive snapshots of zeek logs during ingest
Instead of waiting for all zeek logs to be ready, and then running a
one-time bzng conversion job after that, we now take "snapshots" of
the zeek logs as they are being generated, and make these snapshots
available for regular querying while more logs are being generated.

The word "snapshot" may make this sound fancier than it is - a
snapshot is just the same processing of the zeek logs that we were
doing previously.

Snapshots are taken at exponentially growing intervals, and a final
snapshot is done after all the zeek logs have been crunched.