Rework expression handling a bit
Switch to having evaluators work with native values so we don't
repeatedly decode and encode zvals for complex expressions.  This
also eliminates some of the really repetitive code (although there
is still plenty of repetitive stuff left but such is the nature of
working in a language that doesn't have generics)

At this point, basic numeric expressions work.  Test coverage is not
yet up to par and support for non-numeric types still needs to be added.