add support for pcap-ng (#401)
This commit adds support for pcap-ng.  The pcapio.Reader interface
was tweaked a bit to accommodate the changes so that a Reader can
be either a pcap-ng reader or a legacy pcap reader.  pcapio.Reader
uses the same trick as zio to autodetect whether input is pcap-ng
or pcap legacy.

The pcap-ng parsing code is derived from
https://github.com/google/gopacket/.

We added a simple test for pcap-ng along the lines of the pcap test
that was already there.  Testing is still limited for pcap and
needs to be addressed in a future PR.