Provide better errors on zeek related failures (#405)
Include any output from stderr in error message along with exit status
code and the fact that the error came from the zeek subprocess.

If zeek cannot be found, or the user does not have execute permissions
the request will return with a 500 level errors. All other errors will
occur after the 2xx request headers are sent and thus will be streamed
back in the json TaskEnd.Error payload.