A little more packet post testing (#406)
A consequence of incremental snapshots is that we might have an
all.bzng present in the data dir when aborting. (This wasn't the case
previously, as we renamed all.bzng.tmp to all.bzng a single time,
after successfull completion).

This commit fixes abort() to remove all.bzng upon failure, and adds a
test for this case.

Also,  added checks for new fields in PostStatusMessage, and checks for TaskStart and TaskEnd.