Initial support for expressions (#397)
Add grammar rules for expressions and compilation of expressions.
This is not yet complete (function calls and field/array references
are the big missing pieces).  The plan is to add a `put` proc that can
evaluate expressions (and perhaps to eventually merge some of this
functionality with search filters) but for now the new code is only
referenced from unit tests.

Within the grammar (zql.peg), this patch uses the convention in
many other PEG grammars of rule names written in CamelCase.
this means the grammar is in an inconsistent style right now but the
grammar is overdue for additional clean up, we can make rule names
follow a consistent style as part of that work in a subsequent pass.