Moving zeek process into own package
The zeek pkg introduces the zeek.Launcher func and the zeek.Process
interface. These interfaces allows for test that aren't dependent on
having the zeek binary on the system running the tests.
Additionally passing the pcap reader directly as the zeek process
stdin ensures that we will not get the unfortunate "writing to pipe"
errors that occur when the zeek process is killed during ingest.

For the zqd listen command, if zeek is not available in the system
running the test, disable the packet ingest endpoint and signfy in
the logger output that zeek was not found.