fix offset calculation in legacy pcap reader (#433)
This commit fixes a bug with the pcap reader's offset calculation
at the beginning of the file.  The offset should be adjusted only
when the bytes are actually read instead of when the header is
read into a temporary buffer.  Otherwise, the offset for the pcap
header is incorrect when building a pcap index.  This used to
work because the header offset logic was wired to 0 in the pcap
indexer prior to support for pcap-ng.