fix bug where snaplen = 0
A pcap file with snaplen=0 means that there was no capture length
restriction and all packets are captured at full size.  There was
a check inherited from gopacket that raised an error if the
captured length was larger than the snaplen, which is a bug for
snaplen=0.  The fix is to change this check to check alsoo that
snaplen isn't zero.