make ztest run return separate output and warnings-or-error strings
When ztest run() operates using the zq binary, any warnings generated
will get mixed in with an error message due to the zq command failing.

I expect tests will fall into 3 categories, depending on what they
want to verify: a) just the output, b) output and warnings,
or c) an error message.

This change adds a return value to run() that holds warnings when no
error is reported, or the error string when there is one. This makes it
possible to write tests that fall into one of those 3 categories, _and_
works regardless of whether run() uses the zq binary or evals the zql
programatically with the mux.